OCR/Examples/requirements.txt
astroid==2.3.3
autopep8==1.5
cycler==0.10.0
et-xmlfile==1.0.1
imutils==0.5.3
isort==4.3.21
jdcal==1.4.1
Jinja2==2.11.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
numpy==1.18.1
opencv-python==4.2.0.32
openpyxl==3.0.3
pandas==1.0.0
Pillow==7.0.0
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.6
pytesseract==0.3.2
python-dateutil==2.8.1
pytz==2019.3
scipy==1.4.1
six==1.14.0
tesserocr==2.3.0
typed-ast==1.4.1
Wand==0.5.8
wrapt==1.11.2
